Move day_1 dir -> day_1a.
