+++ advent_of_code_2020/day_1a/sample_input.txt
