+++ advent_of_code_2020/run_test.py
@@ -5,10 +5,10 @@ from click.testing import CliRunner
 from advent_of_code_2020 import run
 
 ROOT_DIR = os.path.dirname(os.path.realpath(__file__))
-INPUT_FILE = os.path.join(ROOT_DIR, "day_1", "sample_input.txt")
+INPUT_FILE = os.path.join(ROOT_DIR, "day_1a", "sample_input.txt")
 
 
-def test_day1() -> None:
+def test_day_1a() -> None:
     runner = CliRunner()
     result = runner.invoke(run.day1, ["--input-file", INPUT_FILE])
     assert result.exit_code == 0, result.output
